GH-262: Optimize Batched Acks
GH-262: In a batch only commit highest offset for each partition  Fixes GH-262 (https://github.com/spring-projects/spring-kafka/issues/262)

This commit adds the highest offset logic to batch listener with non-manual commits. It also adds the logic to the addOffset method to prevent any assumptions from being made.

When dealing with a batch of acks, instead of putting all the consumer records in 'acks' and reducing to the highest offset later (in 'addOffset') reduce the list and add the highest to 'acks'.